requirements.txt
Pillow>=9.5.0
pillow-heif>=0.13.0
tqdm>=4.65.0
customtkinter>=5.2.0
